third commit

+++ utils.py
@@ -5,7 +5,6 @@ import requests
 from bs4 import BeautifulSoup
 
 
-
 def get_top100_list(refresh_html=False):
     """
     실시간 차트 1~100위의 리스트 반환
@@ -16,33 +15,51 @@ def get_top100_list(refresh_html=False):
         51~100위: data/chart_realtime_100.html
     :return:
     """
-    # 프로젝트 컨테이너 폴더 경로
+    # utils가 있는
     path_module = os.path.abspath(__name__)
-    path_module_dir = os.path.dirname(path_module)
-    root_dir = os.path.dirname(os.path.abspath(__name__))
+    print(f'path_module: \n{path_module}')
+
+    # 프로젝트 컨테이너 폴더 경로
+    root_dir = os.path.dirname(path_module)
+    print(f'root_dir: \n{root_dir}')
 
     # data/ 폴더 경로
     path_data_dir = os.path.join(root_dir, 'data')
+    print(f'path_data_dir: \n{path_data_dir}')
 
-
-    # 폴더 만들고 없을경우 에러 나지 않도록.
+    # 만약에 path_data_dir에 해당하는 폴더가 없을 경우 생성해준다
     os.makedirs(path_data_dir, exist_ok=True)
-
 
     # 1~50, 50~100위 웹페이지 주소
     url_chart_realtime_50 = 'https://www.melon.com/chart/index.htm'
     url_chart_realtime_100 = 'https://www.melon.com/chart/index.htm#params%5Bidx%5D=51'
 
-
+    # 1~50위에 해당하는 웹페이지 HTML을
+    # data/chart_realtime_50.html 에 저장
+    # 'xt'모드와 try-except를 쓸 경우
     file_path = os.path.join(path_data_dir, 'chart_realtime_50.html')
-    with open(file_path, 'wt') as f:
-        response = requests.get(url_chart_realtime_50)
-        f.write(response.text)
+    try:
+        with open(file_path, 'xt') as f:
+            response = requests.get(url_chart_realtime_50)
+            source = response.text
+            f.write(source)
+    except FileExistsError:
+        print(f'"{file_path}" file is already exists!')
 
-    file_path2 = os.path.join(path_data_dir, 'chart_realtime_100.html')
-    with open(file_path2, 'wt') as f:
+    # 51~100위에 해당하는 웹페이지 HTML을
+    # data/chart_realtime_100.html 에 저장
+    # 파일이 있는 경우를 검사 후 로직 실행
+    file_path = os.path.join(path_data_dir, 'chart_realtime_100.html')
+    if not os.path.exists(file_path):
         response = requests.get(url_chart_realtime_100)
-        f.write(response.text)
+        source = response.text
+        with open(file_path, 'wt') as f:
+            f.write(source)
+    else:
+        print(f'"{file_path}" file is already exists!')
+
+    # file_path = os.path.join(path_data_dir, 'abc.txt')
+    # print(f'file_path: \n{file_path}')
 
 
 
@@ -141,6 +158,6 @@ def get_top100_list(refresh_html=False):
     #     })
 
 
-# get_top100_list()
+get_top100_list()
 
 print('main.py 실행')
